feat: Refactor authentication handling by centralizing token verification logic and improving middleware checks
